add admin profile view

+++ database/migrations/2023_11_04_015937_create_usuarios_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
 		Schema::create('usuarios', function (Blueprint $table) {
 			$table->id();
 			$table->string('nombre', 255);
-			$table->integer('matricula')->nullable()->unsigned()->unique();
+			$table->bigInteger('matricula')->nullable()->unsigned()->unique();
 			$table->string('correo', 255)->unique();
 			$table->string('password', 255);
 			$table->integer('id_tipo_usuario')->foreign('id_tipo_usuario')->references('id')->on('tipo_usuario');
